pyPurcTally/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pillow
pytesseract
opencv-python
httpx
